requirements.txt
Flask
gunicorn
azure-storage-blob
azure-cosmos
python-dotenv
